+++ src/main/java/com/myZipPlan/server/calculator/dtiCalc/dto/DtiCalcServiceRequest.java
@@ -19,7 +19,7 @@ public class DtiCalcServiceRequest {
     private BigDecimal annualIncome;
     private BigDecimal loanAmount;
     private BigDecimal interestRate;
-    private Integer loanTerm;
+    private BigDecimal loanTerm;
     private RepaymentType repaymentType;
     private BigDecimal yearlyLoanInterestRepayment; // 보유대출 연이자 상환액
