+++ src/main/java/com/myZipPlan/server/calculator/repaymentCalc/domain/RepaymentSchedule.java
@@ -3,17 +3,19 @@ package com.myZipPlan.server.calculator.repaymentCalc.domain;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class RepaymentSchedule {
 
-    private int installmentNumber; // 상환회차
-    private double principalPayment; // 상환원금
-    private double interestPayment; // 상환이자
-    private double totalPayment; // 총 상환액
-    private double remainingPrincipal; // 잔여원금
+    private BigDecimal installmentNumber; // 상환회차
+    private BigDecimal principalPayment; // 상환원금
+    private BigDecimal interestPayment; // 상환이자
+    private BigDecimal totalPayment; // 총 상환액
+    private BigDecimal remainingPrincipal; // 잔여원금
 
     @Builder
-    public RepaymentSchedule(int installmentNumber, double principalPayment, double interestPayment, double totalPayment, double remainingPrincipal) {
+    public RepaymentSchedule(BigDecimal installmentNumber, BigDecimal principalPayment, BigDecimal interestPayment, BigDecimal totalPayment, BigDecimal remainingPrincipal) {
         this.installmentNumber = installmentNumber;
         this.principalPayment = principalPayment;
         this.interestPayment = interestPayment;
