+++ src/main/java/com/myZipPlan/server/calculator/repaymentCalc/dto/RepaymentCalcResponse.java
@@ -4,13 +4,14 @@ import com.myZipPlan.server.calculator.repaymentCalc.domain.RepaymentSchedule;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Builder
 public class RepaymentCalcResponse {
     private List<RepaymentSchedule> repaymentSchedules;
-    private Double totalPrincipal; // 총 원금
-    private Double totalInterest; // 총 이자
-    private Integer totalInstallments; // 총 상환회차
+    private BigDecimal totalPrincipal; // 총 원금
+    private BigDecimal totalInterest; // 총 이자
+    private BigDecimal totalInstallments; // 총 상환회차
 }
